Deleted ReportScreen.tsx and moved the code ResultPage.tsx
